requirements.txt
ariadne==0.13.0
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
graphql-core==3.1.3
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.16.0
redis==3.5.3
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.3
starlette==0.14.2
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
